added interface folder to make a group of the types
