+++ README.md
@@ -1,50 +1,14 @@
-# React + TypeScript + Vite
+# Mirko Saponaro Fullstack Resume
 
-This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.
+Hello! As you can already tell from the title you know my name, 👽.
 
-Currently, two official plugins are available:
+This is an updated version of my previous resume, and I believe it better highlights my growth in designing visually appealing and well-structured webpages.  
+While there’s always room for improvement, I’m continually striving to reach a better level of expertise.
 
-- [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react/README.md) uses [Babel](https://babeljs.io/) for Fast Refresh
-- [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react-swc) uses [SWC](https://swc.rs/) for Fast Refresh
+I’d love the chance to take the next step and discuss the opportunity to join your team.
 
-## Expanding the ESLint configuration
+Looking forward to connecting!  
+Best regards,
 
-If you are developing a production application, we recommend updating the configuration to enable type aware lint rules:
-
-- Configure the top-level `parserOptions` property like this:
-
-```js
-export default tseslint.config({
-  languageOptions: {
-    // other options...
-    parserOptions: {
-      project: ['./tsconfig.node.json', './tsconfig.app.json'],
-      tsconfigRootDir: import.meta.dirname,
-    },
-  },
-})
-```
-
-- Replace `tseslint.configs.recommended` to `tseslint.configs.recommendedTypeChecked` or `tseslint.configs.strictTypeChecked`
-- Optionally add `...tseslint.configs.stylisticTypeChecked`
-- Install [eslint-plugin-react](https://github.com/jsx-eslint/eslint-plugin-react) and update the config:
-
-```js
-// eslint.config.js
-import react from 'eslint-plugin-react'
-
-export default tseslint.config({
-  // Set the react version
-  settings: { react: { version: '18.3' } },
-  plugins: {
-    // Add the react plugin
-    react,
-  },
-  rules: {
-    // other rules...
-    // Enable its recommended rules
-    ...react.configs.recommended.rules,
-    ...react.configs['jsx-runtime'].rules,
-  },
-})
-```
+Mirko Saponaro
+mirko777.au@gmail.com
